Redux

+++ waiwai_backend/waiwai_backend_2/db/migrate/20190914031930_create_restaurants.rb
@@ -5,7 +5,7 @@ class CreateRestaurants < ActiveRecord::Migration[5.2]
       t.text :description
       t.text :image
       t.integer :cost_per_person
-      t.integer :stars
+      t.float :stars
     end
   end
 end
